requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
websockets==12.0
httpx==0.25.2
pydantic==2.5.0
python-dotenv==1.0.0
phidata==2.4.0
openai==1.3.0
pymilvus==2.3.7
